fix: network 모듈 #38
